Added throw exception logic when there is greater demand than available supply. Exceptions are more informative. Moved implementation to separate package. BalanceTransaction now check if Fruit is pre existing in Storage
