+++ src/main/java/core/basesyntax/db/Storage.java
@@ -1,7 +1,6 @@
 package core.basesyntax.db;
 
 import core.basesyntax.model.Fruit;
-
 import java.util.ArrayList;
 import java.util.List;
