+++ src/main/java/core/basesyntax/model/Fruit.java
@@ -22,9 +22,9 @@ public class Fruit {
 
     @Override
     public String toString() {
-        return "Fruit{" +
-                "name='" + name + '\'' +
-                ", quantity=" + quantity +
-                '}';
+        return "Fruit{"
+                + "name='" + name + '\''
+                + ", quantity=" + quantity
+                + '}';
     }
 }
