--- src/main/java/core/basesyntax/service/WriterImpl.java
@@ -1,16 +1,0 @@
-package core.basesyntax.service;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-
-public class WriterImpl implements Writer {
-    @Override
-    public void writeToFile(String fileName, String data) {
-            try {
-                Files.write(Paths.get(fileName), data.getBytes());
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-    }
-}
