+++ src/main/java/core/basesyntax/service/impl/ReportCreator.java
@@ -1,11 +1,10 @@
-package core.basesyntax.service;
-
-import core.basesyntax.model.Fruit;
+package core.basesyntax.service.impl;
 
 import static core.basesyntax.db.Storage.fruits;
+import core.basesyntax.model.Fruit;
 
 public class ReportCreator {
-    public static String createReport(){
+    public static String createReport() {
         StringBuilder report = new StringBuilder();
         report.append("fruit,quantity").append(System.lineSeparator());
         for (Fruit fruit : fruits) {
