+++ src/main/java/core/basesyntax/service/impl/SupplyTransaction.java
@@ -1,14 +1,16 @@
-package core.basesyntax.service;
+package core.basesyntax.service.impl;
+
+import core.basesyntax.service.Transaction;
 
 import static core.basesyntax.db.Storage.fruits;
 
-public class PurchaseTransaction implements Transaction {
+public class SupplyTransaction implements Transaction {
 
     @Override
     public void process(String fruitName, int amount) {
         fruits.stream()
                 .filter(fruit -> fruit.getName().equals(fruitName))
                 .findFirst()
-                .ifPresent(fruit -> fruit.setQuantity(fruit.getQuantity() - amount));
+                .ifPresent(fruit -> fruit.setQuantity(fruit.getQuantity() + amount));
     }
 }
